pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "File diff calculator is a program that compares two files and outputs the result in three styles."
authors = ["Tata Pestova <tata.pestova@inbox.ru>"]
readme = "README.md"
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8.1"
pytest-cov = "^4.0.0"
pyyaml = "^6.0"

[tool.poetry.scripts]
gendiff =  "gendiff.scripts.gendiff:main" 

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.20.2"
flake8 = "^6.0.0"
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
